+++ src/main/java/org/launchcode/codingevents/models/Event.java
@@ -10,10 +10,8 @@ import javax.validation.constraints.Size;
 import java.util.Objects;
 
 @Entity
-public class Event {
-    @Id
-    @GeneratedValue
-    private int id;
+public class Event extends AbstractEntity {
+
 
     @NotBlank(message = "Name is required")
     @Size(min = 3,max = 50,message = "Name must be between 3 and 50 characters")
@@ -92,25 +90,9 @@ public class Event {
         this.type = type;
     }
 
-    public int getId() {
-        return id;
-    }
-
     @Override
     public String toString() {
         return name;
     }
 
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Event event = (Event) o;
-        return id == event.id;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
 }
